grafana/ui: Export types and utils required by grafana/experimental (#42133) (#42220)
(cherry picked from commit d2d38c978fbe94730ac865343b78ed9b02959db9)
