+++ packages/grafana-ui/src/components/Select/Select.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { SelectableValue } from '@grafana/data';
 import { SelectCommonProps, MultiSelectCommonProps, SelectAsyncProps } from './types';
 import { SelectBase } from './SelectBase';
+import { SelectContainer, SelectContainerProps } from './SelectContainer';
 
 export function Select<T>(props: SelectCommonProps<T>) {
   return <SelectBase {...props} />;
@@ -31,3 +32,5 @@ export function AsyncMultiSelect<T>(props: AsyncMultiSelectProps<T>) {
   // @ts-ignore
   return <SelectBase {...props} isMulti />;
 }
+
+export { SelectContainer, SelectContainerProps };
