+++ packages/grafana-ui/src/utils/index.ts
@@ -12,8 +12,12 @@ export { ansicolor } from 'ansicolor';
 
 import * as DOMUtil from './dom'; // includes Element.closest polyfill
 export { DOMUtil };
-export { renderOrCallToRender } from './renderOrCallToRender';
+
 export { createLogger } from './logger';
 export { attachDebugger } from './debug';
 export * from './nodeGraph';
 export { fuzzyMatch } from './fuzzy';
+
+// React utils
+import * as ReactUtils from './reactUtils';
+export { ReactUtils };
