+++ public/app/plugins/datasource/cloudwatch/components/ui/InlineSelect.tsx
@@ -2,7 +2,7 @@ import { css, cx } from '@emotion/css';
 import { GrafanaTheme2 } from '@grafana/data';
 import { Select, stylesFactory, useTheme2 } from '@grafana/ui';
 import {
-  ContainerProps,
+  SelectContainerProps,
   SelectContainer as BaseSelectContainer,
 } from '@grafana/ui/src/components/Select/SelectContainer';
 import { SelectCommonProps } from '@grafana/ui/src/components/Select/types';
@@ -39,7 +39,7 @@ function InlineSelect<T>({ label: labelProp, ...props }: InlineSelectProps<T>) {
 export default InlineSelect;
 
 const SelectContainer = <Option, isMulti extends boolean, Group extends GroupTypeBase<Option>>(
-  props: ContainerProps<Option, isMulti, Group>
+  props: SelectContainerProps<Option, isMulti, Group>
 ) => {
   const { children } = props;
 
@@ -54,7 +54,7 @@ const SelectContainer = <Option, isMulti extends boolean, Group extends GroupTyp
 };
 
 const ValueContainer = <Option, isMulti extends boolean, Group extends GroupTypeBase<Option>>(
-  props: ContainerProps<Option, isMulti, Group>
+  props: SelectContainerProps<Option, isMulti, Group>
 ) => {
   const { className, children } = props;
   const theme = useTheme2();
